feat : resume 페이지 생성 및 라우팅, 기존 레이아웃 변경 (메인페이지의 뱃지 / 자기소개 내용 이동)
